Add a new stubl module for linux 2.6.32, to support Ubuntu Lucid.
This is based on a diff of sit.c between kernels 2.6.25 and 2.6.32.15.
I left out all the network namespace stuff because it doesn't fit well
with our configuration being in module-level parameters.

I haven't tested this with network namespaces, but I'm guessing that
in this situation, the stubl0 interface will be locked to init's namespace,
instead of cloning itself to every namespace like sit.